fixed an issue of outgoing damage buff not being calculated correctly
also readjusted well effect and duration

+++ Assets/Scripts/Ghosts/Specific Ghosts/King/Upgrades/RadientWellEffect.cs
@@ -7,7 +7,7 @@ public class RadientWellEffect : MonoBehaviour
     [SerializeField] float duration;
 
     [Header("Outgoing Damage Percentage Modifiers")]
-    [SerializeField] int damageBuff = 10;
+    [SerializeField] float damageBuff = 10;
 
     [Header("Incoming Damage Percentage Modifiers")]
     [SerializeField] float damageResistance = 10;
